Se agrega update(User user)

+++ src/main/java/ar/com/laboratory/ecommerce/infrastructure/adapter/UserRepositoryImpl.java
@@ -3,8 +3,11 @@ package ar.com.laboratory.ecommerce.infrastructure.adapter;
 import ar.com.laboratory.ecommerce.application.repository.UserRepository;
 import ar.com.laboratory.ecommerce.domain.User;
 import ar.com.laboratory.ecommerce.infrastructure.mapper.UserMapper;
+import ar.com.laboratory.ecommerce.infrastructure.util.exceptions.ProductNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Repository;
+
+import java.util.Objects;
 
 
 @AllArgsConstructor
@@ -13,6 +16,7 @@ public class UserRepositoryImpl implements UserRepository {
 
     private UserCrudRepository repository;
     private UserMapper userMapper;
+    private final static String TABLE = "users";
     @Override
     public Iterable<User> getUsers() {
         return userMapper.toUsers(repository.findAll());
@@ -30,7 +34,14 @@ public class UserRepositoryImpl implements UserRepository {
 
     @Override
     public void deleteUser(Integer id) {
+        var user = repository.findById(id);
+        if (Objects.isNull(user)){
+            throw new ProductNotFoundException(TABLE);
+        }
         repository.deleteById(id);
-
+    }
+    @Override
+    public User update(User user) {
+        return userMapper.toUser(repository.save(userMapper.toEntity(user)));
     }
 }
